Add submission image posting and getting

+++ server.py
@@ -10,7 +10,9 @@ from fastapi import (
     status,
     UploadFile,
     Response,
+    Form
 )
+from dataclasses import dataclass
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from fastapi.staticfiles import StaticFiles
@@ -352,19 +354,35 @@ def get_submission_by_userid(user_id: int):
     return res
 
 
-@app.post("/submissions", tags=["submissions"])
-def post_submission(
-    submission: dto.SubmissionRequest,
-    current_user: Annotated[User, Depends(get_current_user)],
-):
-    service.insert_submission(submission, current_user.id)
-    return {"msg": "Inserted submission"}
+@app.get("/submissions/image", tags=["submissions"])
+def get_submission_image(id: int):
+    res = service.get_submission_image(id)
+    if res is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Not Found.")
+    return res
 
+
+@app.post("/submissions", tags=["submissions"])
+async def post_submission(
+    file: UploadFile,
+    current_user: Annotated[User, Depends(get_current_user)],
+    submission: dto.SubmitForm = Depends()
+):
+    res = await service.insert_submission(submission, current_user.id, file)
+    if (res):
+        return {"msg": "Inserted submission"}
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Not a valid image file, or filesize too big!",
+        )
 
 @app.delete("/submissions", tags=["submissions"])
-def delete_submission(
+async def delete_submission(
     id: int, current_user: Annotated[User, Depends(get_current_user)]
 ):
+    if (not current_user.is_admin):
+        raise auth_err
     service.delete_submission(id)
     return {"msg": "Deleted submission"}
